# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ATBU Hostel Connect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelCampus } from './channelCampus';
import type { ChannelType } from './channelType';

export interface Channel {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  type: ChannelType;
  campus: ChannelCampus;
  /** @nullable */
  hostel?: string | null;
  memberCount?: number;
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt?: string;
}