# lib/api-zod/src/generated/types/sosAlertInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ATBU Hostel Connect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SosAlertInputCampus } from './sosAlertInputCampus';

export interface SosAlertInput {
  userId: string;
  hostel: string;
  campus: SosAlertInputCampus;
  roomNumber?: string;
  message?: string;
}